"Preliminary version 2"

+++ src/reservation/reservation.controller.ts
@@ -1,15 +1,28 @@
-import {Body, Controller, Delete, Get, Param, Post, Query, UseInterceptors, UsePipes} from '@nestjs/common';
+import {
+  BadRequestException,
+  Body,
+  Controller,
+  Delete,
+  ForbiddenException,
+  Get,
+  Param,
+  Post,
+  Request,
+  UseGuards,
+  UseInterceptors,
+  UsePipes,
+} from '@nestjs/common';
 import { HotelRoomService } from '../hotel-room/hotel-room.service';
 import { ReservationDto } from './dto/reservation.dto';
 import { ReservationValidationPipe } from './pipes/reservation-validation.pipe';
 import { ReservationService } from './reservation.service';
-import { CreateReservationInterceptor } from "./interceptors/create-reservation.interceptor";
-import {filter} from "rxjs";
-import {DatesValidationPipe} from "./pipes/dates-validation.pipe";
-import {ID} from "../common/ID";
-import {ReservationListInterceptor} from "./interceptors/reservation-list.interceptor";
-import {Role} from "../roles/role.enum";
-import {Roles} from "../roles/roles.decorator";
+import { CreateReservationInterceptor } from './interceptors/create-reservation.interceptor';
+import { ID } from '../common/ID';
+import { ReservationListInterceptor } from './interceptors/reservation-list.interceptor';
+import { Role } from '../roles/role.enum';
+import { Roles } from '../roles/roles.decorator';
+import { AuthenticatedGuard } from '../auth/guards/authenticated.guard';
+import { RolesGuard } from '../roles/roles.guard';
 
 @Controller()
 export class ReservationController {
@@ -18,39 +31,58 @@ export class ReservationController {
     private readonly reservationService: ReservationService,
   ) {}
 
+  @UseGuards(AuthenticatedGuard, RolesGuard)
   @Post('api/client/reservations')
   @Roles(Role.User) // add roles restriction
   @UseInterceptors(CreateReservationInterceptor) // format output data
   @UsePipes(new ReservationValidationPipe()) // validate input data
-  async createMyReservation(@Body() reservationData: ReservationDto) {
-    reservationData.user = '61e82a837e3bab786f09dabe'; // будем брать из пользователя
-    reservationData.hotel = '61e7f0bf62bc20dd768bf56a'; // будет ясно откуда брать после ответов
-    return await this.reservationService.addReservation(reservationData);
+  async createMyReservation(@Body() data: ReservationDto, @Request() req: any) {
+    // Add user id from req(session)
+    data.user = req.user.id;
+    // Check if room and hotel exist together
+    const roomExist = await this.hotelRoomService.findById(data.room);
+    console.log(roomExist, roomExist.hotelId[0]['_id'].toString());
+    if (!roomExist || roomExist.hotelId[0]['_id'].toString() !== data.hotel) {
+      console.log('room not exist');
+      throw new BadRequestException();
+    }
+    // Try to add reservation:
+    return await this.reservationService.addReservation(data);
   }
-  // 401 - если пользователь не аутентифицирован, 403 - если роль пользователя не client, 400 - если номер с указанным ID не существует или отключен
 
+  @UseGuards(AuthenticatedGuard, RolesGuard)
   @Get('api/client/reservations')
   @Roles(Role.User) // add roles restriction
   @UseInterceptors(ReservationListInterceptor) // format output data
-  myReservationsList() {
-    const user = '61e82a837e3bab786f09dabe';
+  myReservationsList(@Request() req: any) {
+    const user = req.user.id;
     const searchParams = { user };
     return this.reservationService.getReservations(searchParams);
   }
-  // 401 - если пользователь не аутентифицирован
-  // 403 - если роль пользователя не client
 
+  @UseGuards(AuthenticatedGuard, RolesGuard)
   @Delete('api/client/reservations/:id')
   @Roles(Role.User) // add roles restriction
-  async deleteMyReservation(@Param('id') reservationId: ID) {
-    return await this.reservationService.removeReservation(reservationId);
+  async deleteMyReservation(
+    @Param('id') reservationId: ID,
+    @Request() req: any,
+  ) {
+    const reservation = await this.reservationService.getReservations({
+      id: reservationId,
+    });
+    console.log(reservation);
+    if (!reservation[0]) {
+      throw new BadRequestException();
+    }
+    // Compare user and reservation user
+    const userId = req.user.id;
+    if (reservation[0].userId.toString() !== userId) {
+      throw new ForbiddenException();
+    }
+    return this.reservationService.removeReservation(reservationId);
   }
-  // 401 - если пользователь не аутентифицирован
-  // 403 - если роль пользователя не client
-  // 403 - если id текущего пользователя не совпадает с id пользователя в брони
-  // 400 - если бронь с указанным ID не существует
 
-  // MANAGER
+  @UseGuards(AuthenticatedGuard, RolesGuard)
   @Get('api/manager/reservations/:userId')
   @Roles(Role.Manager) // add roles restriction
   @UseInterceptors(ReservationListInterceptor)
@@ -58,19 +90,24 @@ export class ReservationController {
     const searchParams = { user };
     return this.reservationService.getReservations(searchParams);
   }
-  // 401 - если пользователь не аутентифицирован
-  // 403 - если роль пользователя не manager
 
+  @UseGuards(AuthenticatedGuard, RolesGuard)
   @Delete('api/manager/reservations/:userId/:reservationId')
   @Roles(Role.Manager) // add roles restriction
-  async deleteClientReservation(
+  deleteClientReservation(
     @Param('userId') userId: ID,
     @Param('reservationId') reservationId: ID,
   ) {
-    // ПРОВЕРИТЬ, что номер брони и номер пользователя соответствуют друг другу
-    return await this.reservationService.removeReservation(reservationId);
+    const reservation = this.reservationService.getReservations({
+      id: reservationId,
+    });
+    if (!reservation[0]) {
+      throw new BadRequestException();
+    }
+    // Compare user and reservation user
+    if (reservation[0].userId.toString() !== userId) {
+      throw new BadRequestException();
+    }
+    return this.reservationService.removeReservation(reservationId);
   }
-  // 401 - если пользователь не аутентифицирован
-  // 403 - если роль пользователя не manager
-  // 400 - если бронь с указанным ID для пользователя с указанным ID не существует
 }
